% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanModels-data.R
\docType{data}
\name{stan.models}
\alias{stan.models}
\title{stan.models}
\format{An object of class \code{"data.frame"}}
\usage{
stan.models
}
\description{
data.frame containing the structure of the github repository https://github.com/stan-dev/example-models that contains
examples to run STAN models in R from the book by Gelman and Hill 'Data Analysis 
Using Regression Analysis and Multilevel/Hierarchical Models'.
}
\examples{
data(stan.models)
stan.models
}
\keyword{datasets}
