% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3TreeFunctions.R
\name{tree.filter}
\alias{tree.filter}
\title{tree.filter}
\usage{
tree.filter(nodesList, m)
}
\arguments{
\item{m}{data.frame to filter}

\item{nodeList}{list created of tree nodes observed from d3tree.js hook}
}
\value{
data.frame
}
\description{
creates character vector logial expression from tree structure
}
\keyword{internal}
