% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3TreeFunctions.R
\name{tree_filter}
\alias{tree_filter}
\title{tree_filter}
\usage{
tree_filter(nodes_list, m)
}
\arguments{
\item{nodes_list}{list created of tree nodes observed from d3tree.js hook}

\item{m}{data.frame to filter}
}
\value{
data.frame
}
\description{
creates character vector logial expression from tree structure
}
\keyword{internal}
