% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3tree.R
\name{d3tree}
\alias{d3tree}
\title{d3tree}
\usage{
d3tree(data, name = "name", value = "value", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{named list containing hierarchy structure of data created by df2tree and the layout of the tree (collapse,radial,cartesian)}

\item{name}{character containing the names of the nodes}

\item{value}{charater containing the name of the tooltip column
that are used in the leafs}
}
\description{
Htmlwidget that binds to d3js trees. When used in Shiny environment the widget returns
a data.frame of logical expressions that represent the current state of the tree.
}
\examples{
 

\donttest{
if(interactive()){

d3tree(list(root = df2tree(rootname='Titanic',
                           struct=as.data.frame(Titanic)
                           ),
            layout = 'collapse')
      )

d3tree(list(root = df2tree(rootname='Titanic',
                           struct=as.data.frame(Titanic),
                           toolTip=letters[1:(ncol(as.data.frame(Titanic))+1)]),
            layout = 'collapse')
      )
}
}

}

